[FIX] stock_quant_history: port to 17.0
This commit ports the stock_quant_history module to Odoo version 17.0.

The main change is the replacement of the `qty_done` field on the `stock.move.line` model with `quantity`, which is the new field name in Odoo 17.0.

This fixes the following error:
ValueError: Invalid field 'qty_done' on model 'stock.move.line' for 'qty_done:sum'.

The code has also been linted with pre-commit to follow OCA guidelines.